APPLY: allow scrollback using mouse wheel only.
Apply the following patch on top of the previous two to allow scrollback using mouse wheel only when not in MODE_ALTSCREEN. For example the content is being scrolled instead of the scrollback buffer in less. Consequently the Shift modifier for scrolling is not needed anymore. Note: patches before 20191024-a2c479c might break mkeys other than scrolling functions.